refactor: Add modular FastAPI backend with domain-driven structure
Introduce a modular backend structure for the FastAPI project,
splitting authentication, news, and price tracking into dedicated
domain folders. Each domain now contains its own router, models,
schemas, service, config, constants, exceptions, and dependencies
to improve maintainability and scalability. This refactor aligns
with best practices for large-scale FastAPI applications and
prepares the codebase for future feature expansion.
