+++ backend/src/auth/constants.py
@@ -1,0 +1,5 @@
+# Error messages
+USER_ALREADY_EXISTS = "User already exists"
+INVALID_CREDENTIALS = "Invalid credentials"
+USER_NOT_FOUND = "User not found"
+INVALID_TOKEN = "Invalid token"
