+++ backend/src/auth/models.py
@@ -1,0 +1,17 @@
+from sqlalchemy import Column, Integer, String
+from sqlalchemy.orm import relationship
+
+from src.database import Base
+from src.news.models import user_news_association_table
+
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    username = Column(String(50), unique=True, nullable=False)
+    hashed_password = Column(String(200), nullable=False)
+    upvoted_news = relationship(
+        "NewsArticle",
+        secondary=user_news_association_table,
+        back_populates="upvoted_by_users",
+    )
