+++ backend/src/news/constants.py
@@ -1,0 +1,8 @@
+# Error messages
+NEWS_NOT_FOUND = "News article not found"
+INVALID_NEWS_ID = "Invalid news article id"
+AI_SERVICE_ERROR = "AI service error"
+
+# Upvote
+UPVOTE_SUCCESS = "Article upvoted"
+UPVOTE_REMOVED = "Upvote removed"
