+++ backend/src/news/dependencies.py
@@ -1,0 +1,17 @@
+from fastapi import Depends
+from sqlalchemy.orm import Session
+
+from src.database import get_db
+from src.news.service import AIService, NewsService
+
+
+def get_news_service(db: Session = Depends(get_db)) -> NewsService:
+    from src.news.service import ai_service
+
+    return NewsService(db, ai_service)
+
+
+def get_ai_service() -> AIService:
+    from src.news.service import ai_service
+
+    return ai_service
