+++ backend/src/price/constants.py
@@ -1,0 +1,4 @@
+# Error messages
+PRICE_NOT_FOUND = "Price data not found"
+INVALID_PRICE_ID = "Invalid price id"
+PRICE_API_ERROR = "Price API error"
